Create lda.py
This script was used for performing work in a graduate "Research in Composition" course at Texas A&M University - Texarkana. 

Disclaimer: I am a hobbyist, and probably goofed on terminology. My documentation is probably overbearing, and my code is probably itself rather cumbersome. Forgive me, and, if possible, correct me.